# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "btc",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/attempts/submit": {
      "post": {
        "tags": [
          "attempts",
          "attempts"
        ],
        "summary": "Submit Test",
        "operationId": "attempts-submit_test",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitTestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Attempt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/attempts/my": {
      "get": {
        "tags": [
          "attempts",
          "attempts"
        ],
        "summary": "Read User Attempts",
        "description": "Retrieve attempts for the current user.",
        "operationId": "attempts-read_user_attempts",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AttemptPublic"
                  },
                  "title": "Response Attempts-Read User Attempts"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/attempts/stats/me": {
      "get": {
        "tags": [
          "attempts",
          "attempts"
        ],
        "summary": "Get Attempt Stats",
        "operationId": "attempts-get_attempt_stats",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Attempts-Get Attempt Stats"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/attempts/{id}": {
      "get": {
        "tags": [
          "attempts",
          "attempts"
        ],
        "summary": "Read Attempt",
        "operationId": "attempts-read_attempt",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/attempts/export/{test_id}": {
      "get": {
        "tags": [
          "attempts",
          "attempts"
        ],
        "summary": "Export Attempts",
        "operationId": "attempts-export_attempts",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "test_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Test Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Attempt": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "student_id": {
            "type": "string",
            "format": "uuid",
            "title": "Student Id"
          },
          "test_id": {
            "type": "string",
            "format": "uuid",
            "title": "Test Id"
          },
          "score": {
            "type": "integer",
            "title": "Score",
            "default": 0
          },
          "status": {
            "$ref": "#/components/schemas/AttemptStatus",
            "default": "ongoing"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "submitted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "tab_switch_count": {
            "type": "integer",
            "title": "Tab Switch Count",
            "default": 0
          },
          "ai_analysis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Analysis"
          },
          "section_results": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section Results"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "test_id"
        ],
        "title": "Attempt"
      },
      "AttemptAnswerPublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "question_id": {
            "type": "string",
            "format": "uuid",
            "title": "Question Id"
          },
          "selected_option": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Option"
          },
          "answer_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer Text"
          },
          "is_correct": {
            "type": "boolean",
            "title": "Is Correct"
          },
          "marks_obtained": {
            "type": "integer",
            "title": "Marks Obtained"
          },
          "time_spent_seconds": {
            "type": "integer",
            "title": "Time Spent Seconds"
          },
          "question_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Text"
          },
          "solution_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Solution Text"
          },
          "correct_option": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correct Option"
          },
          "correct_answer_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correct Answer Text"
          }
        },
        "type": "object",
        "required": [
          "id",
          "question_id",
          "is_correct",
          "marks_obtained",
          "time_spent_seconds"
        ],
        "title": "AttemptAnswerPublic"
      },
      "AttemptPublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "student_id": {
            "type": "string",
            "format": "uuid",
            "title": "Student Id"
          },
          "test_id": {
            "type": "string",
            "format": "uuid",
            "title": "Test Id"
          },
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "status": {
            "$ref": "#/components/schemas/AttemptStatus"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "submitted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "tab_switch_count": {
            "type": "integer",
            "title": "Tab Switch Count"
          },
          "ai_analysis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Analysis"
          },
          "answers": {
            "items": {
              "$ref": "#/components/schemas/AttemptAnswerPublic"
            },
            "type": "array",
            "title": "Answers",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "student_id",
          "test_id",
          "score",
          "status",
          "started_at",
          "tab_switch_count"
        ],
        "title": "AttemptPublic"
      },
      "AttemptStatus": {
        "type": "string",
        "enum": [
          "ongoing",
          "submitted"
        ],
        "title": "AttemptStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SubmitQuestionResponse": {
        "properties": {
          "question_id": {
            "type": "string",
            "format": "uuid",
            "title": "Question Id"
          },
          "selected_option": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Option"
          },
          "answer_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer Text"
          },
          "time_spent_seconds": {
            "type": "integer",
            "title": "Time Spent Seconds",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "question_id"
        ],
        "title": "SubmitQuestionResponse"
      },
      "SubmitTestRequest": {
        "properties": {
          "test_id": {
            "type": "string",
            "format": "uuid",
            "title": "Test Id"
          },
          "responses": {
            "items": {
              "$ref": "#/components/schemas/SubmitQuestionResponse"
            },
            "type": "array",
            "title": "Responses"
          },
          "tab_switch_count": {
            "type": "integer",
            "title": "Tab Switch Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "test_id",
          "responses"
        ],
        "title": "SubmitTestRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/login/access-token"
          }
        }
      }
    }
  }
}